+++ CHANGELOG.rst
@@ -33,3 +33,4 @@ Added
 - Activation functions for prediction head added.
 - Loss function added.
 - Optimizer function added.
+- Non-Max suppression (NMS) function added